+++ R/03-estimation.R
@@ -172,20 +172,16 @@ ESTIMATION <- function(data, model, group, itemtype = NULL, guess = 0, upper = 1
             data <- remap.distance(data, message = opts$message)
         }
         Data$rowID <- 1L:nrow(data)
-        if(any(rowSums(is.na(data)) == ncol(data))){
-            if(!opts$removeEmptyRows)
-                stop('data contains completely empty response patterns. ',
-                     'Please remove manually or pass removeEmptyRows=TRUE to the technical list',
-                     call.=FALSE)
-            else {
-                pick <- rowSums(is.na(data)) != ncol(data)
-                Data$rowID <- unname(which(pick))
-                data <- subset(data, pick)
-                group <- subset(group, pick)
-                if(!is.null(latent.regression) || !is.null(mixed.design))
-                    stop('removeEmptyRows input not supported for latent regression/mixed effect models. ',
-                         'Please remove the require rows manually for each object.', call.=FALSE)
-            }
+        Data$completely_missing <- which(rowSums(is.na(data)) == ncol(data))
+        if(length(Data$completely_missing)){
+            if(opts$warn)
+                warning('data contains response patterns with only NAs', call.=FALSE)
+            pick <- rowSums(is.na(data)) != ncol(data)
+            Data$rowID <- unname(which(pick))
+            data <- subset(data, pick)
+            group <- subset(group, pick)
+            if(!is.null(latent.regression) || !is.null(mixed.design))
+                covdata <- covdata[-Data$completely_missing, , drop=FALSE]
         }
         Data$data <- data