+++ R/extract.mirt.R
@@ -32,8 +32,7 @@
 #'   \item{itemnames}{a vector of item names from the input data}
 #'   \item{factorNames}{a vector of factor names from the model definition}
 #'   \item{rowID}{an integer vector indicating all valid row numbers used in the model estimation
-#'    (when all cases are used this will be \code{1:nrow(data)}). Mostly useful when the option
-#'     \code{technical = list(removeEmptyRows = TRUE)} was passed}
+#'    (when all cases are used this will be \code{1:nrow(data)})
 #'   \item{data}{raw input data of item responses}
 #'   \item{covdata}{raw input data of data used as covariates}
 #'   \item{tabdatalong}{similar to \code{tabdata}, however the responses have been transformed into
@@ -162,6 +161,7 @@ extract.mirt <- function(x, what){
                   pis = x@Model$pis,
                   nestpars=x@Model$nestpars,
                   prodlist=x@Model$prodlist,
+                  completely_missing=x@Data$completely_missing,
                   stop(sprintf("Could not extract element \'%s\'", what), call.=FALSE))
         ret
 }