+++ R/mirt.R
@@ -595,8 +595,6 @@
 #'       Richardson extrapolation. Default is 1e-5}
 #'     \item{parallel}{logical; use the parallel cluster defined by \code{\link{mirtCluster}}?
 #'       Default is TRUE}
-#'     \item{removeEmptyRows}{logical; remove response vectors that only contain \code{NA}'s?
-#'       Default is FALSE}
 #'     \item{internal_constraints}{logical; include the internal constraints when using certain
 #'       IRT models (e.g., 'grsm' itemtype). Disable this if you want to use special optimizers
 #'       such as the solnp. Default is \code{TRUE}}